modify tablet
